refactor(core): cambiado formato de nombres de tests y colocado Test descolocado en modulo correcto
